requirements.txt
pyperclip
cryptography
pystray
Pillow